src/coordinator.ts/dynamicExport.ts
export function calculateTargetSolarPowerRatio({
    currentSolarWatts,
    targetSolarWatts,
    currentPowerRatio,
}: {
    currentSolarWatts: number;
    targetSolarWatts: number;
    // the current power ratio expressed as a decimal (0.0-1.0)
    currentPowerRatio: number;
}) {
    const estimatedSolarCapacity = currentSolarWatts / currentPowerRatio;
    const targetPowerRatio = targetSolarWatts / estimatedSolarCapacity;

    // cap the target power ratio to 1.0
    return Math.min(targetPowerRatio, 1);
}

// calculate the target solar power to meet the export limit
// note: this may return a value higher than what the PV/inverter is able to produce
// we don't want to make any assumptions about the max capabilities of the inverter due to overclocking
export function calculateTargetSolarWatts({
    solarWatts,
    siteWatts,
    exportLimitWatts,
}: {
    solarWatts: number;
    // the power usage at the site
    // positive = import power
    // negative = export power
    siteWatts: number;
    exportLimitWatts: number;
}) {
    const changeToMeetExportLimit = -siteWatts + -exportLimitWatts;
    const solarTarget = solarWatts - changeToMeetExportLimit;

    return solarTarget;
}
